Add submenu component
